♻️[refactor] 추천 응답 DTO 구조 개선 및 네이밍 통일
